// src/enums/DataEnum.ts
export enum DataEnum {
  DATE = 'RECDATE.DATE',
  TIME = 'RECDATE.TIME',
  MSS_TR = 'MSS_TR',
  TR_UTIL = 'TrUtil',
  TR_UTIL_AVCH = 'TrUtil_AvCH',
  TR_UTIL_HW = 'TrUtil_HW',
  TR_UTIL_SW = 'TrUtil_SW',
  CALL_DROP_RATE = 'CallDropRate',
  C_SET_R = 'CSetR',
  M_ACR = 'mACR',
  S_PAG_R = 'SPagR',
  SLUR = 'SLUR',
  SSMSO = 'SSMSO',
  SSMST = 'SSMST',
  SUSSDR = 'SUSSDR',
  PS_DATA_VOL = 'PS_DATA_VOL',
  Attach_SR_3G = 'Attach_SR_3G',
  Attach_SR_4G = 'Attach_SR_4G',
  PS_Setup_FR_4G = 'PS_Setup_FR_4G',
  PS_Setup_FR_3G = 'PS_Setup_FR_3G',
  SENT_INITIAL_DP_M = 'Sent InitialDP Messages (times)',
  SENT_APPLY_CHARGING_REPORT_M = 'Sent ApplyChargingReport Messages (times)',
  ACH_IDP = 'ACH/IDP',
  PLAY_ANNOUNCMENT_M = 'PlayAnnouncement Messages (times)',
  TC_ABORT_R = 'TC_ABORT_R',
  TC_ABORT_S = 'TC_ABORT_S',
  SENT_TC_ABORT_M = 'Sent TC_ABORT Messages (times)',
  RECEIVED_TC_ABORT_M = 'Received TC_ABORT Messages (times)',
  NUMBER_OF_RECEIVED_PLAY_ANNOUNCMENT_M = 'Number of Received PlayAnnouncement Messages (times)',
  BSC_NAME = 'BSC NAME',
  CDR = 'CDR',
  CunSR = 'CunSR',
  CunSSR = 'CunSSR',
  MOS_2G = 'MOS_2G',
  NUF = 'NUF',
  NUF_G_T = 'NUF_G_T',
  NUF_S_T = 'NUF_S_T',
  NUF_V_T = 'NUF_V_T',
  S_CONG = 'S_CONG',
  S_TRAF = 'S_TRAF',
  SDCDR = 'SDCDR',
  SDR_HO = 'SDR_HO',
  SDR_OTH = 'SDR_OTH',
  SDR_R = 'SDR_R',
  T_CONG = 'T_CONG',
  TAFR = 'TAFR',
  TAFR_NC = 'TAFR_NC',
  TDR_HO = 'TDR_HO',
  TDR_OTH = 'TDR_OTH',
  TDR_R = 'TDR_R',
  TRAF = 'TRAF',
  Natt = 'Natt',
  MHT = 'MHT',
  ASR = 'ASR',
}